검색 시 파라메터 변경 및 수정(셀렉트 박스 부분 및 잘못된 vo 맵핑부분) 및 통계 그래프 색상 변경 및 쿼리 수정

+++ crmai/src/main/java/com/ktds/crmai/model/AIStatistics.java
@@ -1,5 +1,7 @@
 package com.ktds.crmai.model;
 
+import java.math.BigDecimal;
+
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -27,9 +29,9 @@ public class AIStatistics {
 	// 파라메터 방식으로만 사용
 	@Getter @Setter private String beforeDate;			// 기간(이전)
 	@Getter @Setter private String afterDate;			// 기간(이후)
-	@Getter @Setter private Double originalAccAvg;		// 정확도 평균
-	@Getter @Setter private Double soAccAvg;			// 예측 평균
-	@Getter @Setter private Double realAccAvg;			// 결과 평균
+	@Getter @Setter private BigDecimal originalAccAvg;		// 정확도 평균
+	@Getter @Setter private BigDecimal soAccAvg;			// 예측 평균
+	@Getter @Setter private BigDecimal realAccAvg;			// 결과 평균
 	
 	public AIStatistics(String camName, String camStatus, String admId, String camType, String beforeDate,
 			String afterDate) {
